Set "Treat warnings as errors" flag to YES for EmbeddedAudioSynthesis project. Fix all warnings.

+++ common/prboom/p_spec.c
@@ -184,7 +184,7 @@ void P_InitPicAnims (void)
                   animdefs[i].startname,
                   animdefs[i].endname);
 
-    lastanim->speed = LONG(animdefs[i].speed); // killough 5/5/98: add LONG()
+    lastanim->speed = INT32(animdefs[i].speed);
     lastanim++;
   }
   W_UnlockLumpNum(lump);
